# modify ld_prune to prune using correlation block matrix