--- a/python/hail/methods/statgen.py
+++ b/python/hail/methods/statgen.py
@@ -9,11 +9,11 @@ from hail.genetics import KinshipMatrix
 from hail.genetics.reference_genome import reference_genome_type
 from hail.linalg import BlockMatrix
 from hail.matrixtable import MatrixTable
-from hail.methods.misc import require_biallelic, require_col_key_str
+from hail.methods.misc import require_biallelic, require_col_key_str, maximal_independent_set
 from hail.stats import UniformDist, BetaDist, TruncatedBetaDist
 from hail.table import Table
 from hail.typecheck import *
-from hail.utils import wrap_to_list
+from hail.utils import wrap_to_list, new_temp_file
 from hail.utils.java import *
 from hail.utils.misc import check_collisions
 
@@ -2908,15 +2908,74 @@ class FilterAlleles(object):
                 self._left_aligned, self._keep_star))
         return cleanup(m)
 
-
 @typecheck(ds=MatrixTable,
            n_cores=int,
            r2=numeric,
            window=int,
            memory_per_core=int)
 def ld_prune(ds, n_cores, r2=0.2, window=1000000, memory_per_core=256):
-    jmt = Env.hail().methods.LDPrune.apply(ds._jvds, n_cores, r2, window, memory_per_core)
-    return MatrixTable(jmt)
+    """Prune variants in linkage disequilibrium.
+
+    Parameters
+    ----------
+    ds : :class:`.MatrixTable`
+        dataset to prune
+    num_cores : :obj:`int`
+        number of cores to use for local pruning
+    r2 : :obj:`float`
+        correlation threshold (exclusive) above which variants are removed
+    window : :obj:`int`
+        distance in kilobases; correlation between variants further apart is not considered in pruning
+    memory_per_core : :obj:`int`
+        memory in MB per core for local pruning
+
+    Returns
+    -------
+    :class:`.Table`
+        Table of variants after pruning.
+    """
+    sites_only_table = Table(Env.hail().methods.LDPrune.apply(require_biallelic(ds, 'ld_prune')._jvds, 
+                                                       n_cores, float(r2), window, memory_per_core)).persist()
+    locally_pruned_ds = ds.filter_rows(hl.is_defined(sites_only_table[(ds.locus, ds.alleles)]))
+
+    locally_pruned_ds = (locally_pruned_ds.annotate_rows(
+        mean=sites_only_table[(locally_pruned_ds.locus, locally_pruned_ds.alleles)].mean,
+        sd_reciprocal=sites_only_table[(locally_pruned_ds.locus, locally_pruned_ds.alleles)].sd_reciprocal)
+        .add_row_index('row_idx')).persist()
+
+    normalized_mean_imputed_genotype_expr = (
+        hl.cond(hl.is_defined(locally_pruned_ds['GT']),
+                (locally_pruned_ds['GT'].n_alt_alleles() - locally_pruned_ds['mean']) 
+                * locally_pruned_ds['sd_reciprocal'], 0))
+
+    # BlockMatrix.from_entry_expr writes to disk
+    block_matrix = BlockMatrix.from_entry_expr(normalized_mean_imputed_genotype_expr)
+    correlation_matrix = (block_matrix@(block_matrix.T))
+    r2_matrix = correlation_matrix**2
+
+    entries = r2_matrix.entries().persist()
+
+    similar_filter = (entries['entry'] >= r2) & (entries['i'] != entries['j']) & (entries['i'] < entries['j'])
+    entries = entries.filter(similar_filter)
+
+    index_table = locally_pruned_ds.rows().select('locus', 'row_idx').key_by('row_idx')
+    entries = entries.annotate(locus_i=index_table[entries.i].locus, locus_j=index_table[entries.j].locus)
+
+    contig_filter = entries.locus_i.contig == entries.locus_j.contig
+    window_filter = (hl.abs(entries.locus_i.position - entries.locus_j.position)) <= window
+    entries = entries.filter(contig_filter & window_filter)
+
+    if (entries.count()==0):
+        info("LD prune step 3 of 3: nVariantsKept={}".format(locally_pruned_ds.count_rows()))
+        return locally_pruned_ds.rows().select('locus', 'alleles')
+
+    related_nodes_to_remove = maximal_independent_set(entries.i, entries.j, keep=False)
+
+    pruned_ds = locally_pruned_ds.filter_rows(
+        hl.is_defined(related_nodes_to_remove[locally_pruned_ds.row_idx]), keep=False)
+
+    info("LD prune step 3 of 3: nVariantsKept={}".format(pruned_ds.count_rows()))
+    return pruned_ds.rows().select('locus', 'alleles')
 
 
 @typecheck(ds=MatrixTable,