… python/hail/tests/test_methods.py
… src/main/scala/is/hail/methods/LDPrune.scala
… src/main/scala/is/hail/sparkextras/GeneralRDD.scala
… src/main/scala/is/hail/table/Table.scala
… src/test/scala/is/hail/methods/LDPruneSuite.scala
… src/test/scala/is/hail/methods/LocalLDPruneSuite.scala