* modify ld_prune to prune using correlation block matrix
* remove GeneralRDD and globalPrune
* semantic changes to maximalIndpendentSet method 

Context: LDPrune currently does two local pruning steps, followed by a global prune. These changes replace the global pruning step with a method that uses a correlation matrix and computes a maximal independent set of variants from that matrix. The steps are as follows:

1. Local prune in Scala to get a Table with schema {locus, alleles, mean, standard deviation}
2. Get a locally pruned matrix table by filtering the original matrix table using the table from step 1
3. Compute a block matrix from the mean-imputed, standardized genotypes
4. Compute correlation matrix by multiply the block matrix with its transpose, then square the entries of this matrix
5. Get an entries table with schema {i, j, entry} from the r^2 matrix
6. Filter the entries table to the rows where entry>= r^2 threshold, i != j, i < j
7. Annotate entries table with locus (from matrix table in step 2)
8. Filter entries table by contig and window size 
9. Pass the rows of the entries table as edges to the Maximal Independent Set method
10. Filter the locally pruned matrix table to remove the variants returned by MIS, and return this table