@cseed 2018-04-11
Finally, it'd be nice if you can post timings (which I think you have) in the PR: master, vs this branch with a brief description of the timing setup (and maybe the pipeline you're using to time).  Thanks!