@maccum 2018-04-19
benchmarks:

Dataset with 12,000 variants: 

Master: **122 seconds**
6352 variants remain after first local prune 
6329 variants remain after second local prune 
5793 remain after global prune

This branch: **195 seconds** (ran 3 times)
6352 variants remain after local prune
5866 variants remain after final prune

Timing of individual parts for this branch: 
Local prune : ~ 62 seconds
Filtering the matrix table to the locally pruned variants : ~ 20 seconds.
Block matrix of standardized genotypes : ~14 seconds.
Correlation matrix (matrix multiply) : ~67 seconds 
Entries table: ~13 seconds
Annotating entries table, computing maximal independent set, filtering final result: ~17 sec

Dataset with 25,965 variants:

Master : **270 seconds**
12,691 variants (after first local prune) 
12,590 (after second local prune) 
11,749 (after global prune)

This branch:  **503 seconds**
12,691 (after local prune) 
11,851 (after final prune)

Local prune : 83 seconds
Filter matrix table : 37 seconds
Block Matrix: 24 seconds
Correlation Matrix: 265 seconds
Entries Table: 47 seconds
Annotate entries table, MIS, filter after MIS : 44 seconds