@jbloom22 2018-04-21
will be interesting to see how rebase incorporating #3185 affects benchmark