@maccum 2018-04-24
For reference, the commit to "pre-filter entries table for ld_prune" incorporates changes from #3185 to  compute only the blocks from the matrix multiply that are needed in the entries table, rather than computing all blocks. 

503 seconds -> 348 seconds for the 26k variant dataset
195 seconds -> 169 seconds for 12k variant dataset