@cseed 2018-04-29
@jbloom22 While looking over this PR, I ran the following experiment.  Do we have any idea why we're 150x slower than Python here?

```
In [1]: import numpy as np
   ...: x = np.random.rand(1024, 1024)
   ...: x.tofile('x')
   ...: def f():
   ...:   m = np.fromfile('x')
   ...:   n = np.fromfile('x')
   ...:   (n @ (m.T)).tofile('y')
   ...: 
   ...: 

In [2]: %%timeit
   ...: f()
   ...: 
5.95 ms ± 52 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

vs

```
In [1]: import hail as hl
   ...: hl.init()
   ...: def f():
   ...:     m = hl.linalg.BlockMatrix.read('bm')
   ...:     n = hl.linalg.BlockMatrix.read('bm')
   ...:     (m @ (n.T)).write('prod')
   ...: 

In [4]: %%timeit
   ...: f()
   ...: 

887 ms ± 20.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```