@cseed 2018-04-29
It's worse for 4096, >500x: 91.7ms vs 47.9s (!)