@cseed 2018-04-29
And the block matrix in the second case was created with: `hl.linalg.BlockMatrix.random(1024, 1024, 1024).write('bm')`.