@jbloom22 2018-04-29
Oy! If you `grep netlib hail.log`, do you see natives loading or failing to load? I'll check on my end too.