@jbloom22 2018-04-29
We sure seem to be doing the right thing in `build.gradle`: https://github.com/scalanlp/breeze

The current import exists: http://mvnrepository.com/artifact/org.scalanlp/breeze_2.11/0.13.1
```
    compile 'org.scalanlp:breeze-natives_' + scalaMajorVersion + ':' + breezeVersion
```

It looks like Spark 2.2 uses Breeze 0.13.2, search breeze here: https://github.com/apache/spark/blob/branch-2.2/pom.xml

I tried bumping the version, no effect on BLAS though we may as well switch to 0.13.2 once we figure this natives issue out.

I also tried adding this as per the spark docs, though it shouldn't be necessary on OS X. No effect.
```
    compile 'com.github.fommil.netlib:all:1.1.2'
```

Any other ideas?