@cseed 2018-04-29
Yeah, that's it.  Not surprised.

```
2018-04-29 12:41:16 BLAS: WARN: Failed to load implementation from: com.github.fommil.netlib.NativeSystemBLAS
2018-04-29 12:41:16 BLAS: WARN: Failed to load implementation from: com.github.fommil.netlib.NativeRefBLAS
```

I'm on linux.  I'll try to investigate what's going on tonight.