@cseed 2018-04-30
Another thing to explore is how much the block matrix write is spending in compression.  That might not be helping out overall.  It might be worth modifying BlockMatrix to make the compression optional (and add some features from Matrix/Table will make the file format a bit more flexible.)