@cseed 2018-04-30
Also, a quick look at the log shows some AST operations you're using don't have IR conversions, namely: annotate, drop, abs and contains (which @catoverdrive has a PR for: https://github.com/hail-is/hail/pull/3458).  Those would be good additions.